pollen/requirements.txt
Django==3.2.5
djangorestframework==3.12.2
pytest==6.2.4
pytest-django==4.4.0
requests==2.25.1
